@@ -1,8 +1,8 @@
-ruby '2.2.0'
+ruby '2.5.1'
 source 'https://rubygems.org'
 
 # Bundle edge Rails instead: gem 'rails', github: 'rails/rails'
-gem 'rails', '4.1.6'
+gem 'rails', '5.2.0'
 # Use postgresql as the database for Active Record
 gem 'pg'
 # Use SCSS for stylesheets
@@ -10,7 +10,7 @@ gem 'pg'
 # Use Uglifier as compressor for JavaScript assets
 gem 'uglifier', '>= 1.3.0'
 # Use CoffeeScript for .js.coffee assets and views
-gem 'coffee-rails', '~> 4.0.0'
+gem 'coffee-rails', '~> 4.2.2'
 # See https://github.com/sstephenson/execjs#readme for more supported runtimes
 # gem 'therubyracer',  platforms: :ruby
 
@@ -45,7 +45,6 @@ gem 'sass', '~> 3.4.5'
 gem 'haml-rails'
 gem 'compass-rails'
 gem 'google-analytics-rails'
-gem 'rails_config'
 gem 'nokogiri'
 gem 'parallel'
 gem 'kaminari'
@@ -56,6 +55,9 @@ gem 'jquery-ui-rails'
 gem 'underscore-rails'
 gem 'jquery-easing-rails'
 gem 'nprogress-rails'
+gem 'responders'
+gem 'activesupport'
+gem 'config'
 
 # gem 'devise'
 # gem 'redcarpet'# markdown
