@@ -1,4 +1,4 @@
-class AddUrlOriginalToPosts < ActiveRecord::Migration
+class AddUrlOriginalToPosts < ActiveRecord::Migration[4.2]
   def change
     add_column :posts, :url_original, :text
   end
